refactored test file, test fails
