+++ archives_crawler/archives_crawler/spiders/archives_spider_lachlain.py
@@ -1,4 +1,5 @@
 import scrapy
+from spider_utils import gen_keywords
 
 
 class ArchivesSpiderLachlainSpider(scrapy.Spider):
@@ -23,13 +24,5 @@ class ArchivesSpiderLachlainSpider(scrapy.Spider):
                 }
 
 
-    #python / OOP question! Should I have defined this within my class, or outside of it in another class/outside of the class
-    #declaration? Does this matter much?
-    def gen_keywords(fileToOpen):
-        with open(fileToOpen, 'r') as keywordsFile:
-            wordArray = []
-            for line in keywordsFile.readlines():
-                wordArray.append(line)
-                
-            return wordArray
+
