+++ archives_crawler/archives_crawler/spiders/spider_utils.py
@@ -1,0 +1,9 @@
+#python / OOP question! Should I have defined this within my class, or outside of it in another class/outside of the class
+#declaration? Does this matter much?
+def gen_keywords(fileToOpen):
+    with open(fileToOpen, 'r') as keywordsFile:
+        wordArray = []
+        for line in keywordsFile.readlines():
+            wordArray.append(line)
+            
+        return wordArray
