+++ archives_crawler/archives_crawler/spiders/test.txt
@@ -1,0 +1,3 @@
+Kartoffel
+Stein
+Fahrrad
